+++ hal-nano/src/timer.rs
@@ -3,8 +3,8 @@
 //! Uses embassy-time for monotonic microseconds and async delays.
 //! Identical to microRusEFI (same MCU family).
 
-use rusefi_core::hal::{SystemTimer, TimerCallback};
 use embassy_time::{Duration, Instant};
+use rusefi_core::hal::{SystemTimer, TimerCallback};
 
 /// Nano System Timer using embassy-time.
 pub struct Stm32SystemTimer;
@@ -34,7 +34,7 @@ impl Default for Stm32SystemTimer {
 
 impl SystemTimer for Stm32SystemTimer {
     fn now_us(&self) -> u64 {
-        Instant::now().as_micros() as u64
+        Instant::now().as_micros()
     }
 
     fn schedule_us(&mut self, delay_us: u64, callback: TimerCallback) {
