+++ hal-proteus/src/timer.rs
@@ -2,8 +2,8 @@
 //!
 //! Uses embassy-time for monotonic microseconds and async delays.
 
-use rusefi_core::hal::{SystemTimer, TimerCallback};
 use embassy_time::{Duration, Instant};
+use rusefi_core::hal::{SystemTimer, TimerCallback};
 
 /// Proteus System Timer using embassy-time.
 pub struct Stm32SystemTimer;
@@ -33,12 +33,11 @@ impl Default for Stm32SystemTimer {
 
 impl SystemTimer for Stm32SystemTimer {
     fn now_us(&self) -> u64 {
-        Instant::now().as_micros() as u64
+        Instant::now().as_micros()
     }
 
     fn schedule_us(&mut self, delay_us: u64, callback: TimerCallback) {
-        let deadline = embassy_time::Instant::now()
-            + embassy_time::Duration::from_micros(delay_us);
+        let deadline = embassy_time::Instant::now() + embassy_time::Duration::from_micros(delay_us);
         while embassy_time::Instant::now() < deadline {
             core::hint::spin_loop();
         }
